+++ src/components/app.js
@@ -14,12 +14,12 @@ class App extends Component {
           <Menu
             theme="dark"
             mode="horizontal"
-            defaultSelectedKeys={['2']}
+            defaultSelectedKeys={['1']}
             style={{ lineHeight: '64px' }}
           >
-            <Menu.Item key="1">nav 1</Menu.Item>
-            <Menu.Item key="2">nav 2</Menu.Item>
-            <Menu.Item key="3">nav 3</Menu.Item>
+            <Menu.Item key="1">Home</Menu.Item>
+            <Menu.Item key="2">Game</Menu.Item>
+            <Menu.Item key="3">Editor</Menu.Item>
           </Menu>
         </Header>
         <Content style={{ padding: '0 50px' }}>
