+++ src/components/board.js
@@ -2,8 +2,6 @@ import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { bindActionCreators } from 'redux'
 
-import { Row, Col } from 'antd'
-
 import { placeStoneOnBoard } from '../actions/index'
 import BoardNode from './board_node'
 
@@ -11,105 +9,9 @@ class Board extends Component {
   constructor(props) {
     super(props)
 
-    this.state = {
-      prediction: []
-    }
-
     this.handleOnClick = this.handleOnClick.bind(this)
     this.renderRow = this.renderRow.bind(this)
     this.handleCheckTurn = this.handleCheckTurn.bind(this)
-    this.handlePrediction = this.handlePrediction.bind(this)
-    this.drawPrediction = this.drawPrediction.bind(this)
-  }
-
-  drawPrediction() {
-    const style = {
-      backgroundColor: "#CC9966",
-    }
-
-    const rowStyle = {
-      backgroundColor: "#fff",
-      width: "60px",
-      fontWeight: "bold",
-      textAlign: "right",
-      border: "0",
-      paddingRight: "5px",
-    }
-
-    if (!this.props.board.board[0]) {
-      return <div>Loading...</div>
-    }
-
-    return (
-      <table className="table-bordered" style={style}>
-        <thead>
-          <tr>
-            <th></th>
-            {this.props.board.board[0].map((x, i) => <th key={`th${i}`}>{i}</th>)}
-          </tr>
-        </thead>
-        <tbody>
-          {this.props.board.board.map((row, y) => {
-            return (
-              <tr key={y}>
-                <td key={`prediction_td${y}`} style={rowStyle}>{y}</td>
-
-                {row.map((owner, x) => {
-
-                  if (!owner) owner = ""
-
-                  if (this.state.prediction[x] && this.state.prediction[x][y]) {
-                    owner = Math.round(this.state.prediction[x][y])
-                  }
-
-                  const style = {
-                    width: "60px",
-                    height: "60px",
-                    opacity: 1,
-                    borderColor: "#000",
-                    backgroundColor: "#CC9966",
-                    textAlign: "center",
-                  }
-
-                  if (owner >= 2) {
-                    style.backgroundColor = "#fff"
-                  }
-
-                  if (owner === 1) {
-                    style.backgroundColor = "#000"
-                  }
-
-                  return (
-                    <td key={`prediction_${x}-${y}`} style={style}>
-                    </td>
-                  )
-                })}
-              </tr>
-            )
-          })}
-        </tbody>
-      </table>
-    )
-  }
-
-  handlePrediction() {
-    fetch('http://localhost:8081/api/predict', {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'Accept': 'application/json',
-      },
-      body: JSON.stringify({ board: this.props.board.board })
-    })
-      .then(response => response.json())
-      .then(response => {
-        const { prediction } = response
-
-        this.setState({ prediction }, this.drawPrediction)
-      })
-      .catch((e) => {
-        console.log('error in prediction', e)
-      })
   }
 
   handleOnClick(x, y) {
@@ -163,28 +65,17 @@ class Board extends Component {
     }
 
     return (
-      <Row>
-        <Col span="12">
-          <table className="table-bordered" style={style}>
-            <thead>
-              <tr>
-                <th></th>
-                {this.props.board.board[0].map((x, i) => <th key={`th${i}`}>{i}</th>)}
-              </tr>
-            </thead>
-            <tbody>
-              {this.props.board.board.map(this.renderRow)}
-            </tbody>
-          </table>
-        </Col>
-        <Col span="12">
-          {this.drawPrediction()}
-          <br />
-          <button onClick={this.handlePrediction}>
-            Predict
-          </button>
-        </Col>
-      </Row>
+      <table className="table-bordered" style={style}>
+        <thead>
+          <tr>
+            <th></th>
+            {this.props.board.board[0].map((x, i) => <th key={`th${i}`}>{i}</th>)}
+          </tr>
+        </thead>
+        <tbody>
+          {this.props.board.board.map(this.renderRow)}
+        </tbody>
+      </table>
     )
   }
 }
