+++ src/components/game.js
@@ -1,11 +1,27 @@
 import React, { Component } from 'react'
 import { connect } from 'react-redux'
 
+import { Row, Col } from 'antd'
+
 import Board from './board'
+import BoardPrediction from './board_prediction'
 
 class Game extends Component {
   render() {
-    return <Board />
+    return (
+      <div>
+        <Row justify="center">
+          <Col span="12" offset="6">
+            <BoardPrediction />
+          </Col>
+        </Row>
+        <Row>
+          <Col span="12" offset="6">
+            <Board />
+          </Col>
+        </Row>
+      </div>
+    )
   }
 }
