+++ src/components/home.js
@@ -1,0 +1,5 @@
+import React, { Component } from 'react'
+
+const Home = () => <div>Landing page...</div>
+
+export default Home
